# Server/api-spec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.1.0",
        "title": "KFCSmartCorp",
        "description": "HRM System for SmartDEV"
    },
    "host": "localhost:3007",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/leave/leaveEntitlement": {
            "post": {
                "tags": [
                    "Entitlement"
                ],
                "summary": "Create An entitlement",
                "description": "This api use to create new a leave entitlemente\n",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Pet object that needs to be added to the store",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Entitlement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "entitlement created",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/leave/leaveEntitlement/{id}": {
            "delete": {
                "tags": [
                    "Entitlement"
                ],
                "summary": "Delete An entitlement",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteEntitlement",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "The userId that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entitlement delete successful",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/leave/leaveEntitlement/findByEmployeeID/{name}": {
            "get": {
                "tags": [
                    "Entitlement"
                ],
                "summary": "find an entitlement by employeeId",
                "description": "",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "type": "string",
                        "description": "name of employee",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return entitlement",
                        "schema": {
                            "title": "employee object",
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/GetEntitlement"
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/leave/leaveEntitlement/findLeaveTypeByEmployee/{name}": {
            "get": {
                "tags": [
                    "Entitlement"
                ],
                "summary": "find leavetype by employeeId",
                "description": "",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "type": "string",
                        "description": "name of employee",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return leavetype",
                        "schema": {
                            "title": "leaveType object",
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/LeaveTypeForEntitlement"
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/leave/leaveEntitlement/id/{id}/{name}": {
            "get": {
                "description": "",
                "tags": [
                    "Entitlement"
                ],
                "summary": "find entitlement by id and name",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "description": "ID of the employee",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "type": "string",
                        "description": "name of the employee",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return entitlement",
                        "schema": {
                            "title": "employee object",
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/GetEntitlement"
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/leave/leaveEntitlement/findById/{id}": {
            "get": {
                "tags": [
                    "Entitlement"
                ],
                "summary": "find by id",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "description": "id of entitlement",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return entitlement",
                        "schema": {
                            "title": "employee object",
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/GetEntitlement"
                            }
                        }
                    },
                    "401": {
                        "description": "sends request errors",
                        "schema": {
                            "title": "Object",
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Entitlement": {
            "properties": {
                "employee": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/EmployeeForEntitlement"
                    }
                },
                "leaveType": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/LeaveTypeForEntitlement"
                    }
                },
                "entitlementDay": {
                    "type": "string"
                },
                "leavePeriodId": {
                    "type": "string"
                }
            }
        },
        "GetEntitlement": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "employee": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/EmployeeForEntitlement"
                    }
                },
                "leaveType": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/LeaveTypeForEntitlement"
                    }
                },
                "entitlementDay": {
                    "type": "string"
                },
                "leavePeriodId": {
                    "type": "string"
                }
            }
        },
        "LeaveTypeForEntitlement": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "leavetype_id": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                }
            }
        },
        "EmployeeForEntitlement": {
            "properties": {
                "employee_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}